# aulas/aula 06/doc/openapi.yaml
openapi: 3.0.3
info:
  title: Projeto Controle de Filmes - SENAI Jandira
  description: |-
   API destinada ao controle de filmes para o projeto do senai
  termsOfService: http://swagger.io/terms/
  contact:
    email: kauan0alvess@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.4.26
...
servers:
  - url: http://
paths:
...
  /v1/senai/locadora/foto:
    post:
      tags:
        - Foto
      summary: Adiciona uma nova foto
      description: Endpoint para enviar os dados da foto a serem cadastrados
      requestBody:
        description: JSON contendo os dados da nova foto a ser inserido no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Foto'
        required: true
      responses:
        '201':
          description: Success Created Item
        '400':
          description: Error Bad Request
        '415':
          description: Error Content-Type
        '500':
          description: Error Internal Server (Controller / Model)
    get:
      tags:
        - Foto
      summary: Lista todas as fotos
      description: Listagem de todas as fotos registradas no sistema
      responses:
        '200':
          description: Success Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FotoGet'          
        '404':
          description: Error Not Found
        '500':
          description: Error Interal Server (Controller / Model)
  /v1/senai/locadora/foto/{id}:
    get:
      tags:
        - Foto
      summary: Busca uma foto pelo ID
      description: Retorna uma foto pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca da foto
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FotoGet'          
        '400':
          description: Error Bad Request
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller / Model)
    put:
      tags:
        - Foto
      summary: Atualizar uma foto pelo ID
      description: Buscar e atualizar uma foto pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca da foto
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Dados para atualizar uma foto
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Foto'
        required: true
      responses:
        '200':
          description: Success Updated Item
        '400':
          description: Error Bad Request
        '404':
          description: Error Not Found
        '415':
          description: Error Content-Type
        '500':
          description: Error Internal Server (Controller / Model)
    delete:
      tags:
        - Foto
      summary: Excluir uma foto pelo ID
      description: Buscar e excluir uma foto pelo ID
      parameters:
        - name: id
          in: path
          description: ID para buscar a foto
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success Deleted Item
        '400':
          description: Error Bad Request
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller / Model)

  # ENDPOINTS DA TABELA - GÊNERO
  /v1/senai/locadora/genero:
    post:
      tags:
        - Genero
      summary: Adiciona um novo gênero
      description: Endpoint para enviar os dados do gênero a serem cadastrados
      requestBody:
        description: JSON contendo os dados do novo gênero a ser inserido no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Genero'
        required: true
      responses:
        '201':
          description: Success Created Item
        '400':
          description: Error Bad Request
        '415':
          description: Error Content-Type
        '500':
          description: Error Internal Server (Controller / Model)
    get:
      tags:
        - Genero
      summary: Lista todos os gêneros
      description: Listagem de todos os gêneros registrados no sistema
      responses:
        '200':
          description: Success Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GeneroGet'          
        '404':
          description: Error Not Found
        '500':
          description: Error Interal Server (Controller / Model)
  /v1/senai/locadora/genero/{id}:
    get:
      tags:
        - Genero
      summary: Busca um gênero pelo ID
      description: Retorna um gênero pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca do gênero
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneroGet'          
        '400':
          description: Error Bad Request
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller / Model)
    put:
      tags:
        - Genero
      summary: Atualizar um gênero pelo ID
      description: Buscar e atualizar um gênero pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca do gênero
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Dados para atualizar um gênero
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Genero'
        required: true
      responses:
        '200':
          description: Success Updated Item
        '400':
          description: Error Bad Request
        '404':
          description: Error Not Found
        '415':
          description: Error Content-Type
        '500':
          description: Error Internal Server (Controller / Model)
    delete:
      tags:
        - Genero
      summary: Excluir um gênero pelo ID
      description: Buscar e excluir um gênero pelo ID
      parameters:
        - name: id
          in: path
          description: ID para buscar o gênero
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success Deleted Item
        '400':
          description: Error Bad Request
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller / Model)

  # ENDPOINTS DA TABELA - NACIONALIDADE
  /v1/senai/locadora/nacionalidade:
    post:
      tags:
        - Nacionalidade
      summary: Adiciona uma nova nacionalidade
      description: Endpoint para enviar os dados da nacionalidade a serem cadastrados
      requestBody:
        description: JSON contendo os dados da nova nacionalidade a ser inserido no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Nacionalidade'
        required: true
      responses:
        '201':
          description: Success Created Item
        '400':
          description: Error Bad Request
        '415':
          description: Error Content-Type
        '500':
          description: Error Internal Server (Controller / Model)
    get:
      tags:
        - Nacionalidade
      summary: Lista todas as nacionalidades
      description: Listagem de todas as nacionalidades registradas no sistema
      responses:
        '200':
          description: Success Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NacionalidadeGet'          
        '404':
          description: Error Not Found
        '500':
          description: Error Interal Server (Controller / Model)
  /v1/senai/locadora/nacionalidade/{id}:
    get:
      tags:
        - Nacionalidade
      summary: Busca uma nacionalidade pelo ID
      description: Retorna uma nacionalidade pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca da nacionalidade
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NacionalidadeGet'          
        '400':
          description: Error Bad Request
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller / Model) 
    put:
      tags:
        - Nacionalidade
      summary: Atualizar uma nacionalidade pelo ID
      description: Buscar e atualizar uma nacionalidade pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca da nacionalidade
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Dados para atualizar uma nacionalidade
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Nacionalidade'
        required: true
      responses:
        '200':
          description: Success Updated Item
        '400':
          description: Error Bad Request
        '404':
          description: Error Not Found
        '415':
          description: Error Content-Type
        '500':
          description: Error Internal Server (Controller / Model)
    delete:
      tags:
        - Nacionalidade
      summary: Excluir uma nacionalidade pelo ID
      description: Buscar e excluir uma nacionalidade pelo ID
      parameters:
        - name: id
          in: path
          description: ID para buscar a nacionalidade
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success Deleted Item
        '400':
          description: Error Bad Request
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller / Model)

  # ENDPOINTS DA TABELA - SEXO
  /v1/senai/locadora/sexo:
    post:
      tags:
        - Sexo
      summary: Adiciona um novo sexo
      description: Endpoint para enviar os dados do sexo a serem cadastrados
      requestBody:
        description: JSON contendo os dados do novo sexo a ser inserido no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Sexo'
        required: true
      responses:
        '201':
          description: Success Created Item
        '400':
          description: Error Bad Request
        '415':
          description: Error Content-Type
        '500':
          description: Error Internal Server (Controller / Model)
    get:
      tags:
        - Sexo
      summary: Lista todos os sexos
      description: Listagem de todos os sexos registrados no sistema
      responses:
        '200':
          description: Success Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SexoGet'          
        '404':
          description: Error Not Found
        '500':
          description: Error Interal Server (Controller / Model)
  /v1/senai/locadora/sexo/{id}:
    get:
      tags:
        - Sexo
      summary: Busca um sexo pelo ID
      description: Retorna um sexo pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca do sexo
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SexoGet'          
        '400':
          description: Error Bad Request
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller / Model)
    put:
      tags:
        - Sexo
      summary: Atualizar um sexo pelo ID
      description: Buscar e atualizar um sexo pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca do sexo
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Dados para atualizar um sexo
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Sexo'
        required: true
      responses:
        '200':
          description: Success Updated Item
        '400':
          description: Error Bad Request
        '404':
          description: Error Not Found
        '415':
          description: Error Content-Type
        '500':
          description: Error Internal Server (Controller / Model)
    delete:
      tags:
        - Sexo
      summary: Excluir um sexo pelo ID
      description: Buscar e excluir um sexo pelo ID
      parameters:
        - name: id
          in: path
          description: ID para buscar o sexo
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success Deleted Item
        '400':
          description: Error Bad Request
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller / Model)
  # ESTRUTURA DE CADA TABELA:
components:
  schemas:
...
    Foto:
      required: 
        - id
        - foto
      type: object
      properties:
        id:
          type: integer
        foto:
          type: string
          description: O número máximo de caracteres é de 255.
    Genero:
      required:
        - id
        - genero
      type: object
      properties:
        id:
          type: integer
        genero:
          type: string
          example: "Comédia"
          description: O número máximo de caracteres é de 30.
    Nacionalidade:
      required:
        - id
        - nacionalidade
      type: object
      properties:
        id:
          type: integer
        nacionalidade:
          type: string
          example: "Brasileiro"
          description: O número máximo de caraceteres é de 25.
        diretor:
          type: object
          example: {"id": 1}
          description: Coloque , depois de abrir e fechar o json.
        ator:
          type: object
          example: {"id": 1}
          description: Coloque , depois de abrir e fechar o json.
    Sexo:
      required:
        - id
        - sigla
        - sexo
      type: object
      properties:
        id:
          type: integer
        sigla:
          type: string
          example: "M"
          description: O número máximo de caracteres é de 3.
        sexo:
          type: string
          example: "Masculino"
          description: O número máximo de caracteres é de 15.
  
   # SCHEMAS PARA OS GETS (RELAÇÕES)
...
    FotoGet:
      required:
      - id
      - foto
      type: object
      properties:
        id:
          type: integer
        foto:
          type: string
          description: O número máximo de caracteres é de 255.
    GeneroGet:
      required:
      - id
      - genero
      type: object
      properties:
        id:
          type: integer
        genero:
          type: string
          description: O número máximo de caracteres é 100.
        filme:
          type: array
          items:
            $ref: '#/components/schemas/FilmeGeneroGet'
    FilmeGeneroGet:
      required:
      - id
      - nome
      - sinopse
      - capa
      - data_lancamento
      - duracao
      - valor
      - avaliacao
      - id_classificacao
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
          example: "O julgamento da humanidade"
          description: O número máximo de caracteres é 100.
        sinopse:
          type: string
        capa:
          type: string
          description: O número máximo de caracteres é 255.
        data_lancamento:
          type: string
          format: date-time
          example: "2026-05-29"
        duracao:
          type: string
          example: "01:00:00"
        valor:
          type: string
          example: "10.00"
        avaliacao:
          type: string
          example: "5.00"
        id_classificacao:
          type: integer
    NacionalidadeGet:
      required:
      - id
      - nacionalidade
      type: object
      properties:
        id:
          type: integer
        nacionalidade:
          type: string
          description: O número máximo de caracteres é 100.
        diretor:
          type: array
          items:
            $ref: '#/components/schemas/DiretorNacionalidadeGet'
        ator:
          type: array
          items:
            $ref: '#/components/schemas/AtorNacionalidadeGet'
    DiretorNacionalidadeGet:
      required:
      - id
      - nome
      - data_nascimento
      - inicio_carreira
      - id_sexo
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
          example: Caio
          description: O número máximo de caracteres é 100.
        data_nascimento:
          type: string
          format: date-time
          example: "2000-01-01"
        inicio_carreira:
          type: integer
          example: 2020
          description: Ano no formato AAAA (máximo de 4 dígitos).
        id_sexo:
          type: integer
    AtorNacionalidadeGet:
      required:
      - id
      - nome
      - data_nascimento
      - inicio_carreira
      - biografia
      - id_sexo
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
          example: Caio
          description: O número máximo de caracteres é 100.
        data_nascimento:
          type: string
          format: date-time
          example: "2000-01-01"
        inicio_carreira:
          type: integer
          example: 2020
          description: Ano no formato AAAA (máximo de 4 dígitos).
        biografia:
          type: string
        id_sexo:
          type: integer
    SexoGet:
      required:
      - id
      - sigla
      - sexo
      type: object
      properties:
        id:
          type: integer
        sigla:
          type: string
          example: "M"
          description: O número máximo de caracteres é de 3.
        sexo:
          type: string
          example: "Masculino"
          description: O número máximo de caracteres é de 15.
  
  # CAMINHOS PARA DOCUMENTAR
...